small reorg
